==> examples/allganize-rag-evaluation/1_prepare_dataset.ipynb <==
# Prepare Dataset
* apply unique uuid to files (due to issues with fname)

import os
import uuid

import pandas as pd

from config import settings

dataset_dir = os.path.join(settings.data_dir, "retrieval_dataset/allganize-RAG-Evaluation-Dataset-KO")

metadata_df = pd.read_csv(os.path.join(dataset_dir, 'documents.csv'))
print(metadata_df.shape, metadata_df.columns)

answer_df = pd.read_csv(os.path.join(dataset_dir, 'rag_evaluation_result.csv'))
print(answer_df.shape, answer_df.columns)

# 1. DocumentIDs

uuid.uuid4()

metadata_df['id'] = [
    str(uuid.uuid4()) for _ in range(metadata_df.shape[0])
]
metadata_df.to_csv('data/metadata.tsv', sep='\t', index=None)

metadata_df.head()

retrieval_answer_df = answer_df[
    ['domain', 'question', 'target_answer', 'target_file_name', 'target_page_no', 'context_type']
]

retrieval_answer_df.head()

retrieval_answer_df = retrieval_answer_df.merge(
    metadata_df[['id', 'file_name']],
    how='left',
    left_on='target_file_name',
    right_on='file_name'
)

retrieval_answer_df = retrieval_answer_df.drop(columns=['file_name'])
retrieval_answer_df = retrieval_answer_df.rename(
    columns={
        'id': 'target_file_id'
    }
)

retrieval_answer_df = retrieval_answer_df[
    ['domain', 'question', 'target_file_id', 'target_file_name', 'target_page_no', 'context_type', 'target_answer']
]

retrieval_answer_df.head()

retrieval_answer_df.to_csv('data/retrieval_ground_truth.tsv', sep='\t', index=None)
