==> examples/allganize-rag-evaluation/2_read_documents_v2507.ipynb <==
# Read Document Files (v2507)
Read PDF Files into Documents -> Split -> Save to Docstore

## Methodology
```
1. Load Document Readers
    1-1. Initialize DoclingPDFReader (PDF Backend)
    1-2. Initialize DoclingPDFReader (VLM Doctags Backend) - Disabled for now
    1-3. Initialize PDF2ImageReader
2. Load PDF File Data
3. Ingest Data
    3-1. (Reader) PDF File -> PSIKing Document
    3-2. (Splitter) Chunk Documents
4. Insert into DocumentStore
    4-1. Insert
    4-2. Save to Disk
```

import json
import os

from pathlib import Path
import time
from typing import Any, Dict, List, Optional

import pandas as pd
from pydantic import BaseModel
from tqdm import tqdm

from config import settings

from psiking.core.base.schema import Document, TextNode, ImageNode, TableNode

# 1. Load PSIKing Document Readers

## 1-1. Load DoclingPDFReader (PDF Backend)

from docling_core.types.doc import PictureItem

from docling.datamodel.base_models import InputFormat

from docling.datamodel.pipeline_options import (
    AcceleratorDevice,
    VlmPipelineOptions,
    PdfPipelineOptions,
    PictureDescriptionApiOptions,
    ResponseFormat,
    TableStructureOptions,
    TableFormerMode
)
from docling.document_converter import DocumentConverter, PdfFormatOption
from docling.pipeline.vlm_pipeline import VlmPipeline
from docling.datamodel.pipeline_options_vlm_model import (
    ApiVlmOptions,
    InferenceFramework,
    InlineVlmOptions,
    ResponseFormat,
    TransformersModelType
)

pipeline_options = PdfPipelineOptions()

# If force_backend_text = True, text from backend will be used instead of generated text
pipeline_options.force_backend_text = False
pipeline_options.generate_picture_images = True

pipeline_options.images_scale = 1.5
pipeline_options.generate_page_images = True
pipeline_options.generate_picture_images = True
pipeline_options.do_ocr = False

# TableStructure
pipeline_options.do_table_structure = True
pipeline_options.table_structure_options = TableStructureOptions(mode=TableFormerMode.ACCURATE)

pipeline_options.accelerator_options.device = AcceleratorDevice.MPS

from psiking.core.reader.pdf.docling.pipeline_options.picture_description import (
    openai_options as docling_openai_picture_description_options
) 

pipeline_options.do_picture_description = True
pipeline_options.enable_remote_services = True

print(settings.vlm_model)
pipeline_options.picture_description_options=docling_openai_picture_description_options(
    api_key=settings.vlm_api_key,
    model=settings.vlm_model
)

docling_pdf_converter = DocumentConverter(
    allowed_formats = [
        InputFormat.PDF,
    ],
    format_options={
        InputFormat.PDF: PdfFormatOption(
            pipeline_options=pipeline_options,
        ),
    }
)

from psiking.core.reader.pdf.docling import DoclingPDFReader

# initalize reader
docling_pdf_reader = DoclingPDFReader(converter=docling_pdf_converter)

## 1-2. Load DoclingPDFReader (VLM Doctags Backend)
* Disabled for now due to bad Korean ocr performance

# pipeline_options = VlmPipelineOptions()
# # If force_backend_text = True, text from backend will be used instead of generated text
# pipeline_options.force_backend_text = False
# pipeline_options.generate_picture_images = True

# pipeline_options.accelerator_options.device = AcceleratorDevice.MPS


# # smoldocling-preview
# # vlm_conversion_options = InlineVlmOptions(
# #     repo_id = "ds4sd/SmolDocling-256M-preview",
# #     inference_framework=InferenceFramework.TRANSFORMERS,
# #     supported_devices=[AcceleratorDevice.MPS],
# #     transformers_model_type=TransformersModelType.AUTOMODEL_VISION2SEQ,
# #     prompt="Convert this page to docling.",
# #     response_format=ResponseFormat.DOCTAGS,
# #     load_in_8bit=False,
# #     quantized=False
# # )

# ## Remote (vllm) Version
# pipeline_options.enable_remote_services=True
# vlm_conversion_options = ApiVlmOptions(
#     url='http://localhost:8081/v1/chat/completions',
#     prompt="Convert this page to docling.",
#     response_format=ResponseFormat.DOCTAGS,
#     concurrency=8
# )


# ## Pick a VLM model. We choose SmolDocling-256M by default
# pipeline_options.vlm_options = vlm_conversion_options

# docling_vlm_converter = DocumentConverter(
#     format_options={
#         InputFormat.PDF: PdfFormatOption(
#             pipeline_cls=VlmPipeline,
#             pipeline_options=pipeline_options,
#         ),
#     }
# )

# # initalize reader
# docling_vlm_reader = DoclingPDFReader(converter=docling_vlm_converter)

## 1-3. Load PDF2ImageReader

from psiking.core.reader.pdf.pdf2image import PDF2ImageReader

poppler_path = "/opt/homebrew/Cellar/poppler/25.07.0/bin"
pdf2img_reader = PDF2ImageReader(poppler_path=poppler_path)

# 2. Load PDF File Data

# PDF File directory
# pdf_dir = os.path.join(settings.data_dir, "retrieval_dataset/allganize-RAG-Evaluation-Dataset-KO/finance")
pdf_dir = 'data/pdf/finance'
pdf_fnames =[x for x in os.listdir(pdf_dir) if x.endswith(".pdf")]
print("num files:", len(pdf_fnames))
pdf_fnames[:10]

# Map to FileIds
metadata_df = pd.read_csv('data/metadata.tsv', sep='\t')

pdf_file_ids = [
    metadata_df[metadata_df.id==x.replace('.pdf', '')].iloc[0]['id'] for x in pdf_fnames
]

len(pdf_file_ids)

# 3. Ingest Files

## 3-1. Reader - File->Document

# Convert pages to image
documents = []
failed_fnames = []

for doc_i in tqdm(range(len(pdf_fnames))):
    fname=pdf_fnames[doc_i]
    file_path = os.path.join(pdf_dir, fname)
    file_id = pdf_file_ids[doc_i]
    
    # Method 1 - Docling with PDF Backend
    try:
        document = docling_pdf_reader.run(
            file_path, 
            extra_info = {
                "source_id": file_id,
                "domain": "finance",
                "method": "docling-pdf"
            }
        )
        documents.append(document)
        continue
    except Exception as e:
        # print("[DOCLING READER] failed {} - {}".format(fname, str(e)))
        print("[DOCLING PDF READER] failed {}, Falling back to PDF2IMG".format(fname))
        # print(traceback.format_exc())
        
        
    # Method 2 - Docling with VLM (Doctags) Backend (fallback 1)
    ## Disabled due to bad Korean support
    # try:
    #     document = docling_vlm_reader.run(
    #         file_path, 
    #         extra_info = {
    #             "source_id": file_id,
    #             "domain": "finance",
    #             "method": "docling-vlm"
    #         }
    #     )
    #     documents.append(document)
    #     continue
    # except Exception as e:
    #     # print("[DOCLING READER] failed {} - {}".format(fname, str(e)))
    #     print("[DOCLING VLM READER] failed {}, Falling back to PDF2IMG".format(fname))
    #     # print(traceback.format_exc())

    # Method 3 - PDF2Image (fallbakc2)
    try:
        document = pdf2img_reader.run(
            file_path,
            extra_info = {
                "source_id": file_id,
                "domain": "finance",
                "method": "pdf2image"
            }
        )
        documents.append(document)
    except Exception as e:
        print("[PDF2IMG READER] failed {} - {}".format(fname, str(e)))
        failed_fnames.append(fname)

for node in document.nodes[:3]:
    print(type(node))

document.metadata

## 3-2. Splitter - Chunk Documents

from psiking.core.processor.document.text_merger import TextNodeMerger
# Split Documents page-level
merger = TextNodeMerger()

merged_documents = []
for document in documents:
    merged_document = merger.run(document)
    merged_documents.append(merged_document)

# merged_documents[0]
merged_documents[0].nodes[0]

# Run Splitter
import copy
from psiking.core.splitter.text.langchain_text_splitters import LangchainRecursiveCharacterTextSplitter

splitter = LangchainRecursiveCharacterTextSplitter(
    chunk_size = 1024,
    chunk_overlap = 128
)

chunks = []
for document in merged_documents:
    document_chunks = []
    document_metadata = document.metadata
    
    for i, node in enumerate(document.nodes):
        # Run Splitter
        if isinstance(node, TextNode):
            try:
                split_nodes = splitter.run(node)
            except Exception as e:
                print(i, node)
                print(str(e))
                raise e
        else:
            split_nodes = [node]
            
        node_metadata = node.metadata
        # Add 
        chunk_metadata = copy.deepcopy(document_metadata)
        chunk_metadata['prov'] = node_metadata['prov']
        
        # Create New Document
        for split_node in split_nodes:
            # Each Document contains single node
            chunk = Document(
                nodes=[split_node],
                metadata=chunk_metadata
            )
            document_chunks.append(chunk)
    chunks.extend(document_chunks)

print(len(chunks))

# 4. Insert to DocumentStore

from psiking.core.storage.docstore.in_memory import InMemoryDocumentStore

doc_store = InMemoryDocumentStore()

doc_store.add(chunks)

doc_store.count()

doc_store.save('storage/docstore_v2507.json')
