==> examples/allganize-rag-evaluation/3_index_multimodal-hybrid_v2507.ipynb <==
# allganize-RAG-Evaluation data + multimodal hybrid ingestion
## Methodology
```
1. Load Document Readers
    1-1. Load DoclingPDFReader
        1-1-1. Initialize Docling Converter
        1-1-2. Initialize PSIKing Reader
    1-2. Load PDF2ImageReader
2. Load PDF File Data
3. Ingest Data
    3-1. (Reader) PDF File -> PSIKing Document
    3-2. (Splitter) Chunk Documents
4. Embed
5. Insert into DocumentStore, VectorStore
    5-1. Insert to DocStore
    5-2. Insert to VectorStore
6. Test Query
```

## Settings
[Dataset]
* real-life Korean finance pdf files from `allganize-RAG-Evaluation-Dataset-KO`
    * https://huggingface.co/datasets/allganize/RAG-Evaluation-Dataset-KO
    * use 10 'finance' domain files

[Embedder]
* Dense: `jina-embeddings-v4-vllm-retrieval` [[hf link]](https://huggingface.co/jinaai/jina-embeddings-v4-vllm-retrieval)
    * served using vLLM `v0.9.1` docker image
* Sparse: Qdrant/BM42 (`fastembed/sparse/all_miniLM_L6_v2_with_attentions` [[hflink]](https://huggingface.co/Qdrant/all_miniLM_L6_v2_with_attentions))

import json
import os

from pathlib import Path
import time
from typing import Any, Dict, List, Optional

import pandas as pd
from pydantic import BaseModel
from tqdm import tqdm

from config import settings
# Artifacts should contain model weights downloaded using `docling-tools models download`
# Typically set to `~/.cache/docling/models`
# os.environ["DOCLING_ARTIFACTS_PATH"] = settings.docling_artifacts_path

## Import Core Schemas
from psiking.core.base.schema import Document, TextNode, ImageNode, TableNode

# 1. Load Chunks from DocumentStore

## 1-1. Load DocumentStore

from psiking.core.storage.docstore.in_memory import InMemoryDocumentStore

doc_store = InMemoryDocumentStore()

doc_store.load('storage/docstore_v2507.json')

## 1-2. Load Chunks

chunks = list(doc_store._store.values())
print(len(chunks))

chunks[0].metadata

# 4. Embed

## 4-1. Dense Embedding

### 4-1-1. Load Embedder

import asyncio
from tqdm.asyncio import tqdm

from psiking.core.embedder.vllm.online_jina_emb_v4 import VLLMOnlineJinaEmbV4Embedder

dense_embedder = VLLMOnlineJinaEmbV4Embedder(
    base_url=settings.multimodal_embedding_base_url,
    model=settings.multimodal_embedding_model
)

### 4-1-2. Embed Documents

import base64
from io import BytesIO
from PIL import Image

from bs4 import BeautifulSoup

def img_to_base64(img: Image.Image, format="PNG"):
    buffer = BytesIO()
    img.save(buffer, format=format)         # or format="JPEG"
    buffer.seek(0)

    # 3. Base64-encode
    b64 = base64.b64encode(buffer.getvalue()).decode("utf-8")
    return b64

def prepare_text(doc: Document):
    node = doc.nodes[0]
    text = ''
    
    if isinstance(node, TextNode):
        text = node.text
    elif isinstance(node, ImageNode):
        description = node.text
        caption = node.caption
        text = f"Caption: {caption}\nDescription: {description}"
    elif isinstance(node, TableNode):
        html = node.text
        soup = BeautifulSoup(html, 'html.parser')
        table_text = soup.get_text().strip()
        caption = node.caption
        text = f"Caption: {caption}\nTable: {table_text}"
    return text

async def dense_embed(semaphore, doc: Document):
    text = prepare_text(doc)
    messages = [
        {
            'role': 'user',
            'content': [
                {'type': 'text', 'text': 'Passage: '+text},
            ]
        }
    ]
    # Image
    node = doc.nodes[0]
    if isinstance(node, ImageNode) or isinstance(node, TableNode):
        if not node.image_data is None:
            image_base64 = img_to_base64(node.image)
            messages[0]['content'].append(
                {
                    "type": "image_url",
                    "image_url": {"url": f"data:image/png;base64,{image_base64}"}
                }
            )
    async with semaphore:
        try:
            embedding = await dense_embedder.arun(
                input=messages,
                input_format='messages',
                pool=True,
                normalize=True
            )
        except Exception as e:
            print("ERR DOC {} {}".format(doc.id_, str(e)))
            raise e
        
    return embedding

semaphore = asyncio.Semaphore(12)

tasks = []
for chunk in chunks:
    task = dense_embed(semaphore, chunk)
    tasks.append(task)

dense_embeddings = await tqdm.gather(*tasks)

dense_embeddings = [x.tolist() for x in dense_embeddings]

# (num_chunks, seq_len, embedding_dim)
print(len(dense_embeddings))
print(len(dense_embeddings[0]))

type(dense_embeddings[0][0])

## 4-2. Sparse Embedding

### 4-2-1. Load Embedder

from fastembed import SparseTextEmbedding

sparse_model_dir = os.path.join(
    settings.model_weight_dir, "embedding/fastembed/sparse/all_miniLM_L6_v2_with_attentions"
)
os.listdir(sparse_model_dir)

sparse_embedding_model = SparseTextEmbedding(
    model_name="Qdrant/bm42-all-minilm-l6-v2-attentions",
    specific_model_path=sparse_model_dir,
    cuda=False,
    lazy_load=False
)

from psiking.core.embedder.fastembed.local_sparse import LocalFastEmbedSparseEmbedder

sparse_embedder = LocalFastEmbedSparseEmbedder(
    model=sparse_embedding_model
)

### 4-2-2. Embed Documents

sparse_inputs = [
    prepare_text(x) for x in chunks
]

sparse_embedding_values, sparse_embedding_indices = sparse_embedder.run(
    sparse_inputs,
    batch_size=256
)

# 5. Insert into DocumentStore, VectorStore

## 5-2. Insert to VectorStore

from qdrant_client import QdrantClient
from psiking.core.storage.vectorstore.qdrant import QdrantSingleHybridVectorStore

# initialize client
# client = QdrantClient(":memory:")
client = QdrantClient(host="localhost", port=6333)
collection_name = "allganize-finance-multimodal-hybrid-v2507_2"

vector_store = QdrantSingleHybridVectorStore(
    collection_name=collection_name,
    client=client
)

from qdrant_client.http import models

dense_embedding_dim = len(dense_embeddings[0])

vector_store.create_collection(
    on_disk_payload=True,  # store the payload on disk
    dense_vector_config = models.VectorParams(
        size=dense_embedding_dim,
        distance=models.Distance.COSINE,
        on_disk=True,
        hnsw_config = {
            "m": 16,
            "ef_construct": 100,
        }
    ),
    sparse_vector_config = models.SparseVectorParams(
        modifier=models.Modifier.IDF, ## uses indices from bm42 embedder
    )
)

# check collection
collection_info = vector_store._client.get_collection(
    collection_name=vector_store.collection_name
)
print(collection_info.model_dump_json(indent=4))

vector_store.add(
    documents=chunks,
    dense_embeddings=dense_embeddings,
    sparse_embedding_values=sparse_embedding_values,
    sparse_embedding_indices=sparse_embedding_indices,
    metadata_keys=["source_id", "domain", 'prov', 'method']
)

chunks[0].id_

points = vector_store._client.retrieve(
    collection_name=vector_store.collection_name,
    ids=[chunks[0].id_],
    with_vectors=True
)

print(points[0].id)
print(points[0].payload)
print(len(points[0].vector))

# 6. Test Query

import numpy as np
from psiking.core.storage.vectorstore.schema import (
    MetadataFilters,
    FilterOperator,
    VectorStoreQuery,
    VectorStoreQueryMode,
    VectorStoreQueryOptions,
)   

# Use random query embedding
query_embedding = np.random.randn(dense_embedding_dim)

vsquery=VectorStoreQuery(
    dense_embedding=query_embedding,
    sparse_embedding_values=[],
    sparse_embedding_indicies=[]
    
)
vsoptions=VectorStoreQueryOptions(
    mode=VectorStoreQueryMode.HYBRID,
    top_k=10,
    hybrid_fusion_method='rrf',
    sparse_top_k=30,
    dense_top_k=30
)

points = vector_store.query(
    query=vsquery,
    options=vsoptions
)

points

# Get Retrieved Result from docstore
retrieved_doc_id = points[0].id

retrieved_doc = doc_store.get(retrieved_doc_id)[0]

nodes = retrieved_doc.nodes
print(len(nodes))
print(json.dumps(retrieved_doc.metadata, indent=4))
