==> examples/allganize-rag-evaluation/4_evaluate-retrieval_text_v2507.ipynb <==
# allganize-RAG-Evaluation - retrieval performance
## Methodology
```
1. Load DocStore, VectorStore
    1-1. Load DocStore
    1-2. Load VectorStore
2. Initialize Embedder
3. Load Evaluation Data
    3-1. Load Query & Ground Truth
    3-2. Calculate Query Embeddings
4. Run Retrieval
```


import json
import os

from config import settings

INDEXING_MODE="text-hybrid"

# 1. Load DocStore, VectorStore

## 1-1. Load DocStore

from psiking.core.storage.docstore.in_memory import InMemoryDocumentStore

doc_store = InMemoryDocumentStore()
doc_store.load('storage/docstore_v2507.json')

## 1-2. Load VectorStore

from qdrant_client import QdrantClient
from psiking.core.storage.vectorstore.qdrant import QdrantSingleVectorStore

# initialize client
# client = QdrantClient(":memory:")
client = QdrantClient(host="localhost", port=6333)
collection_name = f"allganize-finance-{INDEXING_MODE}-v2507_2"
# collection_name = f"allganize-finance-{INDEXING_MODE}-v2507"

vector_store = QdrantSingleVectorStore(
    collection_name=collection_name,
    client=client
)

# 2. Intialize Embedder

## 2-1. Dense Embedder

import asyncio
from tqdm.asyncio import tqdm

from psiking.core.embedder.vllm.online_jina_emb_v4 import VLLMOnlineJinaEmbV4Embedder

dense_embedder = VLLMOnlineJinaEmbV4Embedder(
    base_url=settings.multimodal_embedding_base_url,
    model=settings.multimodal_embedding_model
)

## 2-2. Sparse Embedder

from fastembed import SparseTextEmbedding
from psiking.core.embedder.fastembed.local_sparse import LocalFastEmbedSparseEmbedder

sparse_model_dir = os.path.join(
    settings.model_weight_dir, "embedding/fastembed/sparse/all_miniLM_L6_v2_with_attentions"
)
os.listdir(sparse_model_dir)

sparse_embedding_model = SparseTextEmbedding(
    model_name="Qdrant/bm42-all-minilm-l6-v2-attentions",
    specific_model_path=sparse_model_dir,
    cuda=False,
    lazy_load=False
)

sparse_embedder = LocalFastEmbedSparseEmbedder(
    model=sparse_embedding_model
)

# 3. Load Evaluation Data

## 3-1. Load Query & Ground Truth

import pandas as pd
answer_df = pd.read_csv('data/retrieval_ground_truth.tsv', sep='\t')
print(answer_df.shape[0])
answer_df = answer_df[answer_df.domain=='finance']
print(answer_df.shape[0])

answer_df.head()

## 3-2. Calculate Query Embeddings

import asyncio
from typing import List
from tqdm.asyncio import tqdm as atqdm

import numpy as np
from psiking.core.storage.vectorstore.schema import (
    MetadataFilters,
    FilterOperator,
    VectorStoreQuery,
    VectorStoreQueryMode,
    VectorStoreQueryOptions,
)   

queries = answer_df.question.values.tolist()

### 3-2-1. Calculate Dense Embeddings

async def calculate_dense_embedding(semaphore, text: str, mode: str='messages'):
    if mode=='messages':
        messages = [
            {
                'role': 'user',
                'content': [
                    {'type': 'text', 'text': text},
                ]
            }
        ]
        async with semaphore:
            try:
                embedding = await dense_embedder.arun(
                    input=messages,
                    input_format='messages',
                    pool=True,
                    normalize=True
                )
            except Exception as e:
                print("ERR {}".format( str(e)))
                raise e
    else:
        async with semaphore:
            try:
                embedding = await dense_embedder.arun(
                    input=text,
                    input_format='text',
                    mode='query',
                    pool=True,
                    normalize=True
                )
            except Exception as e:
                print("ERR {}".format( str(e)))
                raise e
        
    return embedding

QUERY_EMBEDDING_MODE='text'
QUERY_EMBEDDING_MODE='messages'

semaphore = asyncio.Semaphore(16)
tasks = []
for query in queries:
    task = calculate_dense_embedding(semaphore, query, mode=QUERY_EMBEDDING_MODE)
    tasks.append(task)

dense_query_embeddings = await tqdm.gather(*tasks)

### 3-2-2. Calculate Sparse Embeddings

sparse_query_embedding_values, sparse_query_embedding_indicies = sparse_embedder.run(
    queries,
    batch_size=256
)

sparse_query_embedding_values[0], sparse_query_embedding_indicies[0]

# 4. Prepare Evaluation

from typing import Literal
## Retrieval Function
def retrieve(
    mode: Literal['dense', 'sparse', 'hybrid'],
    d: np.ndarray,
    s_v: List[float],
    s_i: List[float],
    k: int = 10,
    sparse_k: int = 30,
    dense_k: int = 30,
)->List[str]:
    """Retrieves point(chunk's) documentids"""
    vsquery=VectorStoreQuery(
        dense_embedding=d,
        sparse_embedding_values=s_v,
        sparse_embedding_indicies=s_i
    )
    if mode=='hybrid':
        vsoptions=VectorStoreQueryOptions(
            mode=VectorStoreQueryMode.HYBRID,
            top_k=k,
            hybrid_fusion_method='rrf',
            sparse_top_k=sparse_k,
            dense_top_k=dense_k
        )
    elif mode=='dense':
        vsoptions=VectorStoreQueryOptions(
            mode=VectorStoreQueryMode.DENSE,
            top_k=k,
        )
    
    elif mode=='sparse':
        vsoptions=VectorStoreQueryOptions(
            mode=VectorStoreQueryMode.SPARSE,
            top_k=k,
        )
    
    chunks = vector_store.query(
        query=vsquery,
        options=vsoptions
    )
    chunk_ids = [x.id for x in chunks]
    return chunk_ids

## Evaluation Function
from utils.eval_utils import (
    calculate_filelevel_ap,
    calculate_filelevel_rr,
    calculate_pagelevel_ap,
    calculate_pagelevel_rr
)

# 5. Evaluate - Sparse

k = 30

retrieved_chunkids = []

print("K {}".format(k))
for query_i in tqdm(range(len(queries))):
    d = dense_query_embeddings[query_i]
    s_v = sparse_query_embedding_values[query_i]
    s_i = sparse_query_embedding_indicies[query_i]
    
    retrieved_chunkids.append(
        retrieve(
            'sparse',
            d=d,
            s_v=s_v,
            s_i=s_i,
            k=k,
        )
    )

## 5-1. File-Level

answer_fileids = [
    answer_df.iloc[x]['target_file_id']
    for x in range(len(queries))
]

retrieved_fileids = [
    [
        doc_store.get(x)[0].metadata['source_id']
        for x in chunkids
    ]
    for chunkids in retrieved_chunkids
]
len(answer_fileids), len(retrieved_fileids[5])

# [len(x) for x in retrieved_fileids]

# mAP
print(QUERY_EMBEDDING_MODE)
for at_k in [5,10, 15]:
    aps = [
        calculate_filelevel_ap(retrieved, answer, at_k=at_k)
        for retrieved, answer in zip(retrieved_fileids, answer_fileids, strict=True)
    ]
    print("mAP@{}: {:.4f}".format(at_k, sum(aps)/len(aps)))

# mRR
print(QUERY_EMBEDDING_MODE)
for at_k in [5,10,15]:
    rrs = [
        calculate_filelevel_rr(retrieved, answer, at_k=at_k)
        for retrieved, answer in zip(retrieved_fileids, answer_fileids, strict=True)
    ]

    print("mRR@{}: {:.4f}".format(at_k, sum(rrs)/len(rrs)))

## 5-2. Page-Level

answers = [
    (answer_df.iloc[x]['target_file_id'], int(answer_df.iloc[x]['target_page_no']))
    for x in range(len(queries))
]

retrieved_metadata = [
    [
        doc_store.get(x)[0].metadata
        for x in chunkids
    ]
    for chunkids in retrieved_chunkids
]

print(QUERY_EMBEDDING_MODE)
for at_k in [5,10,15]:
    aps = [
        calculate_pagelevel_ap(retrieved, answer, at_k=at_k)
        for retrieved, answer in zip(retrieved_metadata, answers, strict=True)
    ]
    print("mAP@{}: {:.4f}".format(at_k, sum(aps)/len(aps)))

print(QUERY_EMBEDDING_MODE)
for at_k in [5,10,15]:
    rrs = [
        calculate_pagelevel_rr(retrieved, answer, at_k=at_k)
        for retrieved, answer in zip(retrieved_metadata, answers, strict=True)
    ]
    print("mRR@{}: {:.4f}".format(at_k, sum(rrs)/len(rrs)))

# 6. Evaluate - Dense

k = 30

retrieved_chunkids = []

print("K {}".format(k))
for query_i in tqdm(range(len(queries))):
    d = dense_query_embeddings[query_i]
    s_v = sparse_query_embedding_values[query_i]
    s_i = sparse_query_embedding_indicies[query_i]
    
    retrieved_chunkids.append(
        retrieve(
            'dense',
            d=d,
            s_v=s_v,
            s_i=s_i,
            k=k,
        )
    )

## 6-1. File-Level

answer_fileids = [
    answer_df.iloc[x]['target_file_id']
    for x in range(len(queries))
]

retrieved_fileids = [
    [
        doc_store.get(x)[0].metadata['source_id']
        for x in chunkids
    ]
    for chunkids in retrieved_chunkids
]

# mAP
print(QUERY_EMBEDDING_MODE)
for at_k in [5,10,15]:
    aps = [
        calculate_filelevel_ap(retrieved, answer, at_k=at_k)
        for retrieved, answer in zip(retrieved_fileids, answer_fileids, strict=True)
    ]
    print("mAP@{}: {:.4f}".format(at_k, sum(aps)/len(aps)))

# mRR
print(QUERY_EMBEDDING_MODE)
at_k = 5
for at_k in [5,10,15]:
    rrs = [
        calculate_filelevel_rr(retrieved, answer, at_k=at_k)
        for retrieved, answer in zip(retrieved_fileids, answer_fileids, strict=True)
    ]

    print("mRR@{}: {:.4f}".format(at_k, sum(rrs)/len(rrs)))

## 6-2. Page-Level

answers = [
    (answer_df.iloc[x]['target_file_id'], int(answer_df.iloc[x]['target_page_no']))
    for x in range(len(queries))
]

retrieved_metadata = [
    [
        doc_store.get(x)[0].metadata
        for x in chunkids
    ]
    for chunkids in retrieved_chunkids
]

print(QUERY_EMBEDDING_MODE)
# mAP
for at_k in [5,10,15]:
    aps = [
        calculate_pagelevel_ap(retrieved, answer, at_k=at_k)
        for retrieved, answer in zip(retrieved_metadata, answers, strict=True)
    ]
    print("mAP@{}: {:.4f}".format(at_k, sum(aps)/len(aps)))

print(QUERY_EMBEDDING_MODE)
# mRR
for at_k in [5,10,15]:
    rrs = [
        calculate_pagelevel_rr(retrieved, answer, at_k=at_k)
        for retrieved, answer in zip(retrieved_metadata, answers, strict=True)
    ]
    print("mRR@{}: {:.4f}".format(at_k, sum(rrs)/len(rrs)))

# 7 - Evaluate - Hybrid

k = 30
sparse_k = 100
dense_k = 100

retrieved_chunkids = []

print("K {} (sparse {} dense {})".format(k, sparse_k, dense_k))
for query_i in tqdm(range(len(queries))):
    d = dense_query_embeddings[query_i]
    s_v = sparse_query_embedding_values[query_i]
    s_i = sparse_query_embedding_indicies[query_i]
    
    retrieved_chunkids.append(
        retrieve(
            'hybrid',
            d=d,
            s_v=s_v,
            s_i=s_i,
            k=k,
            sparse_k=sparse_k,
            dense_k=dense_k
        )
    )

## 7-2. File-Level

answer_fileids = [
    answer_df.iloc[x]['target_file_id']
    for x in range(len(queries))
]

retrieved_fileids = [
    [
        doc_store.get(x)[0].metadata['source_id']
        for x in chunkids
    ]
    for chunkids in retrieved_chunkids
]

print(QUERY_EMBEDDING_MODE)
# mAP
for at_k in [5,10,15]:
    aps = [
        calculate_filelevel_ap(retrieved, answer, at_k=at_k)
        for retrieved, answer in zip(retrieved_fileids, answer_fileids, strict=True)
    ]
    print("mAP@{}: {:.4f}".format(at_k, sum(aps)/len(aps)))

print(QUERY_EMBEDDING_MODE)
# mRR
at_k = 5
for at_k in [5,10,15]:
    rrs = [
        calculate_filelevel_rr(retrieved, answer, at_k=at_k)
        for retrieved, answer in zip(retrieved_fileids, answer_fileids, strict=True)
    ]

    print("mRR@{}: {:.4f}".format(at_k, sum(rrs)/len(rrs)))

## 7-2. Page Level

answers = [
    (answer_df.iloc[x]['target_file_id'], int(answer_df.iloc[x]['target_page_no']))
    for x in range(len(queries))
]

retrieved_metadata = [
    [
        doc_store.get(x)[0].metadata
        for x in chunkids
    ]
    for chunkids in retrieved_chunkids
]

print(QUERY_EMBEDDING_MODE)
# mAP
for at_k in [5,10,15]:
    aps = [
        calculate_pagelevel_ap(retrieved, answer, at_k=at_k)
        for retrieved, answer in zip(retrieved_metadata, answers, strict=True)
    ]
    print("mAP@{}: {:.4f}".format(at_k, sum(aps)/len(aps)))

print(QUERY_EMBEDDING_MODE)
# mRR
for at_k in [5,10,15]:
    rrs = [
        calculate_pagelevel_rr(retrieved, answer, at_k=at_k)
        for retrieved, answer in zip(retrieved_metadata, answers, strict=True)
    ]
    print("mRR@{}: {:.4f}".format(at_k, sum(rrs)/len(rrs)))
